# api/swagger.yaml
openapi: 3.0.3
info:
  title: Medical Information System (MIS) - Task Tracker API
  description: |-
    API модуля трекера рабочих задач для медицинского персонала (врачи, администраторы, медсестры).
  version: 1.0.0
servers:
  -
    url: http://localhost:8080/api/v1
    description: Локальный сервер разработки (Docker)

paths:
  /tasks:
    post:
      summary: Создать новую рабочую задачу
      description: Создает задачу в системе (одиночную или периодический шаблон). По умолчанию новой задаче присваивается статус `NEW`. Принимает список имен тегов.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Задача успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Некорректные входные данные или дата выполнения в прошлом
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера

    get:
      summary: Получить список задач с фильтрацией и пагинацией
      description: Возвращает отсортированный по дате выполнения (due_date) список обычных и виртуальных повторений периодических задач, соответствующих фильтрам и параметрам пагинации за указанный период. Каждая задача возвращается со своими тегами.
      parameters:
        -
          name: status
          in: query
          description: Фильтр по текущему статусу задачи
          required: false
          schema:
            $ref: '#/components/schemas/TaskStatus'
        -
          name: due_date_from
          in: query
          description: Нижняя граница даты выполнения (формат RFC3339)
          required: false
          schema:
            type: string
            format: date-time
            example: '2026-06-18T00:00:00Z'
        -
          name: due_date_to
          in: query
          description: Верхняя граница даты выполнения (формат RFC3339)
          required: false
          schema:
            type: string
            format: date-time
            example: '2026-06-19T23:59:59Z'
        -
          name: page
          in: query
          description: Номер запрашиваемой страницы (по умолчанию 1)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        -
          name: limit
          in: query
          description: Количество элементов на страницу (по умолчанию 20, максимум 100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
      responses:
        '200':
          description: Страница списка задач успешно получена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTasksResponse'
        '500':
          description: Внутренняя ошибка сервера

  /tasks/{id}:
    get:
      summary: Получить детальную информацию о задаче по ID
      description: Возвращает полные данные задачи, включая список привязанных тегов с их метаданными.
      parameters:
        -
          name: id
          in: path
          required: true
          description: Идентификатор задачи
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Данные задачи успешно получены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Некорректный формат ID задачи
        '404':
          description: Задача с указанным ID не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Полное обновление существующей задачи
      description: Обновляет поля задачи, параметры её периодичности и полностью перезаписывает Many-to-Many связи с тегами. Несуществующие в справочнике теги тихо пропускаются (Silent Fail).
      parameters:
        -
          name: id
          in: path
          required: true
          description: Идентификатор задачи
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Задача успешно обновлена
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: updated
        '400':
          description: Некорректный JSON или невалидные данные
        '404':
          description: Задача не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Удалить задачу из системы
      parameters:
        -
          name: id
          in: path
          required: true
          description: Идентификатор задачи
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Задача успешно удалена (нет содержимого в ответе)
        '400':
          description: Некорректный формат ID
        '404':
          description: Задача не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /tasks/{id}/executions:
    post:
      summary: Отметить статус выполнения конкретного календарного дня задачи
      description: Изолированное действие персонала МИС для фиксации факта выполнения периодической задачи на определенную дату без изменения основного шаблона.
      parameters:
        -
          name: id
          in: path
          required: true
          description: Идентификатор периодической задачи-шаблона
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordExecutionRequest'
      responses:
        '200':
          description: Статус дня успешно зафиксирован
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: execution_recorded
        '400':
          description: Некорректный JSON или невалидные параметры даты/статуса
        '500':
          description: Внутренняя ошибка сервера

  /tags:
    get:
      summary: Получить справочник тегов МИС с пагинацией
      description: Возвращает список доступных тегов (включая системные) с поддержкой постраничного вывода для инициализации интерфейсов МИС.
      parameters:
        -
          name: page
          in: query
          description: Номер запрашиваемой страницы (по умолчанию 1)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        -
          name: limit
          in: query
          description: Количество элементов на страницу (по умолчанию 20, максимум 100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
      responses:
        '200':
          description: Успешный возврат страницы справочника тегов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTagsResponse'
        '500':
          description: Внутренняя ошибка сервера

    post:
      summary: Создать новый кастомный тег в справочнике
      description: Явное добавление нового тега администратором МИС. По умолчанию создается с флагом `is_system = false`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagCreateRequest'
      responses:
        '201':
          description: Тег успешно создан в справочнике
        '400':
          description: Невалидные входные данные или некорректный JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера

  /tags/{name}:
    put:
      summary: Изменить имя существующего кастомного тега
      description: Переименовывает тег в справочнике МИС. Связанные задачи каскадно обновляются силами СУБД (ON UPDATE CASCADE). Изменение системных тегов запрещено.
      parameters:
        -
          name: name
          in: path
          required: true
          description: Текущее уникальное имя тега для изменения
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagUpdateRequest'
      responses:
        '200':
          description: Тег успешно переименован
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: updated
        '400':
          description: Невалидное новое имя или некорректный JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Попытка изменения защищенного системного тега
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Указанный тег не найден в справочнике
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера

    delete:
      summary: Удалить кастомный тег из системы
      description: Удаляет тег из справочника МИС. Связи с задачами вычищаются каскадно на уровне БД (ON DELETE CASCADE). Удаление системных тегов под запретом.
      parameters:
        -
          name: name
          in: path
          required: true
          description: Уникальное имя тега для удаления
          schema:
            type: string
      responses:
        '204':
          description: Тег успешно удален из системы
        '403':
          description: Попытка удаления защищенного системного тега
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера

components:
  schemas:
    TaskStatus:
      type: string
      enum: [NEW, IN_PROGRESS, DONE, CANCELED]
      example: NEW

    CreateTaskRequest:
      type: object
      required:
        - title
        - due_date
      properties:
        title:
          type: string
          description: Краткое название медицинской задачи
          example: 'Сделать обход пациентов'
        description:
          type: string
          description: Подробное описание действий по задаче
          example: 'Проверить пульс и давление у пациентов в палатах 3 и 4.'
        due_date:
          type: string
          format: date-time
          description: Планируемое время задачи / Старт расписания для периодических задач
          example: '2026-12-31T15:00:00Z'
        tags:
          type: array
          description: Список имен тегов, выбранных на форме создания
          items:
            type: string
          example: ['отчетность', 'операции']
        recurrence:
          $ref: '#/components/schemas/RecurrenceDTO'

    UpdateTaskRequest:
      type: object
      required:
        - title
        - due_date
        - status
      properties:
        title:
          type: string
          example: 'Сделать обход пациентов (Срочно)'
        description:
          type: string
          example: 'Проверить пульс, давление и выдать новые назначения.'
        due_date:
          type: string
          format: date-time
          example: '2026-12-31T16:00:00Z'
        status:
          $ref: '#/components/schemas/TaskStatus'
        tags:
          type: array
          description: Список имен тегов, выбранных на форме редактирования (перетирает старые связи)
          items:
            type: string
          example: ['операции', 'звонок']
        recurrence:
          $ref: '#/components/schemas/RecurrenceDTO'

    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: 'Сделать обход пациентов'
        description:
          type: string
          example: 'Проверить пульс и давление у пациентов в палатах 3 и 4.'
        due_date:
          type: string
          format: date-time
          example: '2026-12-31T15:00:00Z'
        status:
          $ref: '#/components/schemas/TaskStatus'
        tags:
          type: array
          description: Список объектов тегов с их метаданными
          items:
            $ref: '#/components/schemas/TagStr'
        is_recurring:
          type: boolean
          description: Флаг периодичности для отрисовки иконок повторения на фронтенде МИС
          example: true

    RecurrenceDTO:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum: [DAILY, MONTHLY, DATES, EVEN, ODD]
          example: DAILY
        interval:
          type: integer
          description: Интервал повторения в днях (для типа DAILY)
          example: 1
        day_of_month:
          type: integer
          description: Число месяца (для типа MONTHLY, от 1 до 31)
          example: 15
        specifics:
          type: array
          description: Набор конкретных изолированных дат (для типа DATES)
          items:
            type: string
            format: date-time
          example: ['2026-07-10T00:00:00Z', '2026-07-15T00:00:00Z']

    RecordExecutionRequest:
      type: object
      required:
        - date
        - status
      properties:
        date:
          type: string
          format: date-time
          description: Конкретная календарная дата выполнения виртуального экземпляра задачи
          example: '2026-06-22T00:00:00Z'
        status:
          $ref: '#/components/schemas/TaskStatus'

    TagCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 50
          example: 'консилиум'

    TagUpdateRequest:
      type: object
      required:
        - new_name
      properties:
        new_name:
          type: string
          maxLength: 50
          example: 'Оториноларинголог'

    TagResponse:
      type: object
      properties:
        name:
          type: string
          example: 'отчетность'
        is_system:
          type: boolean
          example: true

    TagStr:
      type: string
      example: "Лор"

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: 'task not found'

    PaginationMetadata:
      type: object
      required:
        - current_page
        - limit
        - total_items
        - total_pages
      properties:
        current_page:
          type: integer
          description: Номер текущей страницы
          example: 1
        limit:
          type: integer
          description: Количество элементов на одной странице
          example: 20
        total_items:
          type: integer
          description: Общее количество элементов во всей базе данных по заданному фильтру
          example: 142
        total_pages:
          type: integer
          description: Общее количество доступных страниц
          example: 8

    PaginatedTasksResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          description: Массив задач на текущей странице
          items:
            $ref: '#/components/schemas/TaskResponse'
        pagination:
          $ref: '#/components/schemas/PaginationMetadata'

    PaginatedTagsResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          description: Массив тегов на текущей странице
          items:
            $ref: '#/components/schemas/TagResponse'
        pagination:
          $ref: '#/components/schemas/PaginationMetadata'